Phase3 :: Restructed the code with reusable components
