+++ src/config/navigation.tsx
@@ -1,0 +1,21 @@
+import AccountCircleIcon from "@mui/icons-material/AccountCircle";
+import ListIcon from "@mui/icons-material/List";
+import MapIcon from "@mui/icons-material/Map";
+
+export const navigationItems = [
+  {
+    text: "Map",
+    icon: <MapIcon />,
+    path: "/",
+  },
+  {
+    text: "List",
+    icon: <ListIcon />,
+    path: "/list",
+  },
+  {
+    text: "Account",
+    icon: <AccountCircleIcon />,
+    path: "/account",
+  },
+];
