+++ src/data/facilities.ts
@@ -1,0 +1,31 @@
+import { Facility } from "../types/facility";
+
+export const facilities: Facility[] = [
+  {
+    id: 1,
+    name: "Downtown Office",
+    address: "123 Main Street, New York, NY 10001",
+    icon: "🏢",
+    assets: 50000,
+    lat: 40.7128,
+    lng: -74.006,
+  },
+  {
+    id: 2,
+    name: "West Coast Hub",
+    address: "456 Tech Boulevard, San Francisco, CA 94105",
+    icon: "🏭",
+    assets: 85000,
+    lat: 37.7749,
+    lng: -122.4194,
+  },
+  {
+    id: 3,
+    name: "South Center",
+    address: "789 Palm Avenue, Miami, FL 33101",
+    icon: "🏗️",
+    assets: 75000,
+    lat: 25.7617,
+    lng: -80.1918,
+  },
+];
