+++ src/types/facility.ts
@@ -1,0 +1,9 @@
+export interface Facility {
+  id: number;
+  name: string;
+  address: string;
+  icon: string;
+  assets: number;
+  lat: number;
+  lng: number;
+}
